#373 move net_question functions to nextquestion.c
